Split FINUFFT backends onto two extensions
This is needed if we want to use the CPU-based FINUFFTBackend without
explicitly loading CUDA.
