+++ Project.toml
@@ -35,7 +35,8 @@ Observables = "510215fc-4207-5dde-b226-833fc4488ee2"
 ThreadPinning = "811555cd-349b-4f26-b7bc-1f208b848042"
 
 [extensions]
-VortexPastaFINUFFTExt = ["FINUFFT", "CUDA"]
+VortexPastaCuFINUFFTExt = ["FINUFFT", "CUDA"]
+VortexPastaFINUFFTExt = ["FINUFFT"]
 VortexPastaMakieExt = ["MakieCore", "Observables"]
 VortexPastaThreadPinningExt = ["ThreadPinning"]
